Move classificator class and common things to separate files
