+++ CMakeLists.txt
@@ -4,6 +4,13 @@ project( boolean_calculator_cli )
 
 set( CMAKE_CXX_STANDARD 20 )
 
+# TODO: Separate library from executable to custom functions
+# (no need to build CLI for Frontend)
+
+add_library( classificator
+    src/classificator.cpp src/classificator.hpp
+)
+
 add_library( base_input_manager
     src/base_input_manager.cpp src/base_input_manager.hpp
 )
@@ -26,6 +33,10 @@ target_link_libraries( cli_input_manager
 
 add_library( cli_output_manager
     src/cli_output_manager.cpp src/cli_output_manager.hpp src/base_output_manager.hpp
+)
+
+target_link_libraries( cli_output_manager
+    classificator
 )
 
 add_library( cli_lib
