+++ src/cli_output_manager.hpp
@@ -2,6 +2,7 @@
 #define CLI_OUTPUT_MANAGER_HPP
 
 #include "base_output_manager.hpp"
+#include "classificator.hpp"
 
 #include <iostream>
